requirements.txt
pygame~=2.6.1
requests~=2.32.3
mysql-connector-python